a2a-plaground/server-agent/pyproject.toml
[project]
name = "server-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk[a2a]>=1.6.1",
]
